hw: i386: Factorize fw_cfg initialisation code
The fw_cfg device initialisation code is moved into its own file and
made PC machine type agnostic.
We need to do this code factorization in order for all x86 machine types
to be able initialize and add a fw_cfg device, including the new virt
one.
